refactor(sketching): invert OO/Fns for CompoundSketch
Layer 3 audit F5 (CompoundSketch portion) — same inversion as PR #955
applied to CompoundSketch. The shell-generator helpers (faceFromWires,
guessFaceFromWires, fixWire, solidFromShellGenerator) and all four
operations (face, extrude, revolve, loftWith) move into sketchFns.ts.
Class methods become 1-line delegations.

Behaviour unchanged.
